+++ custom_components/wattpilot/switch.py
@@ -119,7 +119,7 @@ class ChargerSwitch(ChargerPlatformEntity):
     async def async_turn_on(self, **kwargs) -> None:
         """Async: Turn entity on"""
         try:
-            _LOGGER.debug("%s - %s: async_turn_on: %s", self._charger_id, self._identifier, self._attr_name)
+            _LOGGER.debug("%s - %s: async_turn_on: %s", self._charger_id, self._identifier, self._attr_translation_key)
             if self._entity_cfg.get('invert', False):
                 value = False
             else:
@@ -132,7 +132,7 @@ class ChargerSwitch(ChargerPlatformEntity):
     async def async_turn_off(self, **kwargs) -> None:
         """Async: Turn entity off"""
         try:
-            _LOGGER.debug("%s - %s: async_turn_off: %s", self._charger_id, self._identifier, self._attr_name)
+            _LOGGER.debug("%s - %s: async_turn_off: %s", self._charger_id, self._identifier, self._attr_translation_key)
             if self._entity_cfg.get('invert', False):
                 value = True
             else: